refactor: refactor `get_ffa_feature_test_target`
Change `get_ffa_feature_test_target` to return a `size_t` instead of an
`unsigned int`, because `size_t` is the return type of operators like
`sizeof()`.

Change `get_ffa_feature_test_target` to require its argument to be
non-null (and assert that it is). This function is only used for getting
the array of features to test, so there is no use case where passing a
non-null pointer would make sense.

+++ include/runtime_services/spm_common.h
@@ -103,10 +103,10 @@ void dump_ffa_value(struct ffa_value ret);
 
 bool check_spmc_execution_level(void);
 
-unsigned int get_ffa_feature_test_target(
-		const struct ffa_features_test **test_target);
+size_t get_ffa_feature_test_target(
+	const struct ffa_features_test **test_target);
 bool ffa_features_test_targets(const struct ffa_features_test *targets,
-			       uint32_t test_target_size);
+			       size_t test_target_size);
 
 /**
  * Helper to conduct a memory retrieve. This is to be called by the receiver
